coproc/tests: Move router_test_fixture into its own file
